Restructure provisioning files
Add new 'provision' folder.

Move files related to after a submit endpoint (a message sent to
the submit queue) into the 'provision' folder for
organizational/logical purposes. Provision will be restructured
to support partial provisioning and the number of submit endpoints
will grow. To avoid crowding of the 'portfolioDrafts/submit' folder
an additional folder was added under 'packages/api'.

Ticket: AT-6844
